feat: Reality
- Pico providers registered through Monika can set their names through
the `MaSetPicoProviderName` function.
- Added a custom WSL device, `/dev/reality`. WSL1 Pico processes can
switch to other providers through an `ioctl` on that device. Reading
this device gives information about the current providers for the
current process along with Monika driver details. Writing to the device
is undefined behavior (mostly equivalent to writing to `/dev/null` but
with some Easter eggs).

The `ioctl` function is untested, so expect BSODs when using it.