chore: Document newly added field
The last field, `SubsystemInformationType`, was not available in the
earliest versions of Windows (the ones where thinkcz's Pico toolbox was
tested on).

The field is used by `NtQueryInformationProcess` when the query is of
class `ProcessSubsystemInformation`. The `1` value corresponds to
`SubsystemInformationTypeWSL`.

With information about this field, and further knowledge about the
`OpenProcess` and `OpenThread` access masks, more complete checks have
now been implemented in the heuristics for
`PicoSppLocateProviderRoutines`.