chore: Add pre-calculated offsets
Integrated with the `dist` branch of
https://github.com/trungnt2910/PspPicoProviderRoutinesOffsetGen/.

The offsets in the above mentioned repo are updated weekly. However,
here in `lxmonika` the files are still manually synchronized.

This addition allows more reliable findings of the
`PspPicoProviderRoutines` struct in more versions of Windows, including
Windows 10 and the some of the latest Windows 11 Insider builds. It also
includes offsets for ARM64 builds, simplifying the future process of
porting the driver the ARM64.

The offsets also include all the functions needed to synthesize our own
`PS_PICO_ROUTINES` struct, allowing `lxmonika` to theoretically work
even on machines with WSL disabled.