chore: Huge overhaul of lxmonika semantics
- Pico process and thread context objects are now managed by lxmonika.
The driver transparently boxes and unboxes the pointers while
intercepting Pico callbacks.
- Provider information is now stored in the context object. The previous
process map implementation has been removed, not only because a map is
less efficient than directly storing pointers on the corresponding NT
objects, but also because the whole idea of one provider info for a
whole process is broken.
- When a provider calls the terminate process/thread callback, if the
process/thread has a parent provider (in other words, the current driver
did not start the process/thread, but was handed control through the
reality device), lxmonika will no longer forward the call to Windows,
but instead allow the call to return and sends the object to the parent
provider. This allows any potential cleanup callbacks from the parent to
be properly called.
- WSL (lxcore.sys) is now pulled in as a lxmonika provider, if it is
enabled on the system. The subsystem will always have a provider ID of
0. This allows a unified process/thread lifetime management model.
- `/dev/reality` is now process-specific, not thread. The `ioctl`
operation will only affect the current thread. The `read` operation will
return thread-specific data, and will therefore return different results
for different threads even when using the same FD in the same process.

This commit does come with some regressions:
- `/dev/reality` files will no longer be cached. The buffer is
regenerated on every `read` operation. However, this should not be a
major issue, since this device is seldom accessed except by system
information scripts or for debugging purposes.
- Our killer feature, the WSL1 `uname` patch, does not work anymore.
Currently, callbacks to and from WSL1 receive no special treatment,
making it hard for the dispatcher to single out Linux syscalls. This
will be fixed in an upcoming commit.