fix: Restore WSL syscall hook
Restores the `uname` syscall hook by overriding the provider entry for
`lxcore` with a special function.

The pre/post-syscall design in the previous versions is removed. We do
not intend to expose syscall hooking mechanism to other drivers, so
it is best to keep the syscall pipeline as simple as possible.

`uname` in WSL now displays `Monix` once again.