fix: Fall back to process for DispatchException
Falls back to getting the context and provider index associated with
the current process if doing so for the current thread fails.

This prevents a loop of exceptions being dispatched to the original
routine (usually `lxcore`) when a non-Pico thread does something weird
in a Pico process. This happens, for example, when Task Manager tries
to attach a thread to a process.