docs: Update driver installation instructions
Prefer using the `sc` command to register lxmonika as a driver
service instead of using `devcon.exe`. Benefits of this method
includes:
- The external `devcon.exe` tool is not required. `sc.exe` is a
Windows built-in utility.
- No dummy device needs to be registered.
- Easy and clean uninstallation.

The only drawback of this method is that the installed drivers
cannot start on boot for some reason
(see https://github.com/jephthai/OpenPasswordFilter/issues/1 for
a similar issue). However, we can still set the "system" start type,
which is still early enough to hook `lxcore`.