chore(host): Remove Switch::Validate
It does not make much sense to include the function in the Switch
class.

Unlike the initial design, the `Switch` class is often constructed
inline during the constructor of the owning `Command` class.

Normally, `Switch` will not be inherited. Instead, the `Parameter`
class should be inherited, also providing some general validation.
Deeper validation should occur in the command itself.

We cannot move this `Validate` method to `Parameter` since
`Parameter` does not know the type of the target variable before
`Parse` gets called.