refactor(host): Move Commands to its own directory
Since we're adding a lot of `Command`s soon, it makes sense to
have a dedicated directory.

We might not do the same for `Switch`es since they're all created
inline in the `Command`s' constructors, but we might want to do the
same when the number of `Parameter`s get too large to fit in
`Parameter.cpp`.