refactor(mxss): Unify arch-specific syscall logic
Use the same macro to implement syscalls for all supported
architectures.

Each arch has one instruction, one reg for the syscall number, 6 regs
for the arguments, one reg for the syscall return value (which may be
the same as the one for the syscall number or one of the arguments).
The macro also accepts an optional list of clobbered registers or
resources in addition to the default "memory".

This also somewhat demystifies the GCC extended inline asm syntax.
"r" can be used for all registers and "r+" means the register will be
written into. The first line is the instruction, the second line are
output operands, the third line are inputs, and the fourth line is a
list of clobbered registers/resources.

The declarations have been checked with the implementation in `musl`
and tested on `x86` and `x86_64`. For `arm64` and `arm`, Compiler
Explorer shows the expected assembly.