fix(host): Fix uninstall command
`monika uninstall` will now make sure that no more dependent services
are running before allowing itself to be removed.

It will also try to delete `lxmonika.sys` from the system drivers
directory.

The command is still somewhat unsafe at this stage, especially if WSL
is involved. It is still recommended to reboot as early as possible.