fix: Correctly document and handle ABI break
While the structure ABI break affecting AoW is at RS1, the signature
break is not there until RS4.

This commit renames everything to RS4 to reflect this discovery.
It also implements a method to differentiate between older and newer
versions of the function, since we cannot rely on structure sizes to
differentiate between RS1 and RS4.