fix(mxss): Identify correct ABI version
The "modern" Pico ABI was previously incorrectly labeled as RS1, while
in fact the signature changes occurred at RS4.

This commit fixes `mxss` to account for the changes.
`mxss`, unlike the rest of `lxmonika`, is designed to only build with
the latest WDK and target the latest versions of Windows. It therefore
identifies with the version of the WDK it was built, but at least RS4
to prevent random bugs should we ever decide to retarget the driver.