feat(pico): Directly use `PsRegisterPicoProvider`
Add support for running `lxmonika.sys` as a core driver.

In this process, `lxmonika` would try to detect the sizes of kernel
Pico structures and register its own compatible callbacks.

If the driver is loaded later, the size detection code will still work,
but registration will be done using the legacy offset and patching
method.

If loaded as a core driver, the new `lxmonika` will be incompatible
with the `monika` CLI.