fix: Guard DriverEntry
Add a guard allowing `DriverEntry` to only run once.

This is added to support the `lxmonika.sys` driver being registered
twice: Once injected as a "Core" driver service to handle Pico
functionality, and once as itself (i.e. `lxmonika`) for ease of
dependency management in the `monika.exe` CLI tool.