chore(module): Add hooking routines
Add a `MdlpPatchImport` function that patches the IAT of a certain
module. This function is known to NOT work on core Windows drivers that
bypass the IAT, such as `lxcore.sys`.

To work around this, this commit also adds a function
`MdlpPatchTrampoline` that directly installs a `jmp` trampoline to hook
the function. This approach is dirtier and architecture-specific but
does the job.