feat: Initial support for 32-bit platforms
lxmonika now builds on x86 and ARM32, with the target version lowered
to Win10 2004.

A lot of hacks are involved to get this unsupported version of WDK to
build on Visual Studio 2022. We rely on the latest `.props` and
`.targets` to do the job for us, just that the binaries and headers
are for an older version of Windows.

An attempt was made to further lower the bar to RS1, but it turned out
that we used too many Win10 2004+ functions so we will stop here for
now.

None of the built binaries have been tested.