feat: Device sets
Add "device sets", a mechanism for `lxmonika` to deal with the fact that
it needs, and will need, to run through multiple driver initialization
funtions and share its driver object with foreign libraries.

Devices under the `lxmonika` object (or whatever object `lxmonika`
pretends to be) will have their major functions hidden behind a
common set of dispatcher functions that redirects according to the
active device object, based on the "device set" it belongs to.